Correct Conversion from JsonNode
