+++ kondor-core/src/main/kotlin/com/ubertob/kondor/json/JJsonNode.kt
@@ -1,27 +1,15 @@
 package com.ubertob.kondor.json
 
 import com.ubertob.kondor.json.JsonStyle.Companion.appendNode
-import com.ubertob.kondor.json.jsonnode.*
+import com.ubertob.kondor.json.jsonnode.JsonNode
+import com.ubertob.kondor.json.jsonnode.JsonNodeNull
+import com.ubertob.kondor.json.jsonnode.JsonNodeObject
+import com.ubertob.kondor.json.jsonnode.ObjectNode
 
-object JJsonNode : ObjectNodeConverterProperties<JsonNodeObject>() {
+object JJsonNode : JAny<JsonNodeObject>() {
     override val _nodeType = ObjectNode
     override fun toJsonNode(value: JsonNodeObject): JsonNodeObject =
         value
-
-    override fun fromFieldValues(fieldValues: FieldsValues, path: NodePath): JsonOutcome<JsonNodeObject> =
-        tryFromNode(path) {
-            val nodeMap = fieldValues.mapValues { value -> //!!! duplicated
-                when (value) {
-                    null -> JsonNodeNull
-                    is String -> JsonNodeString(value)
-                    is Number -> JsonNodeNumber(value)
-                    is Boolean -> JsonNodeBoolean(value)
-                    is JsonNode -> value
-                    else -> throw JsonParsingException(ConverterJsonError(path, "Unsupported type: ${value::class}"))
-                }
-            }
-            JsonNodeObject.buildForParsing(nodeMap, path)
-        }
 
     override fun fieldAppenders(valueObject: JsonNodeObject): List<NamedAppender> =
         valueObject._fieldMap.map
@@ -35,4 +23,6 @@ object JJsonNode : ObjectNodeConverterProperties<JsonNodeObject>() {
             null
         else
             { style, off -> appendNode(node, style, off) }
+
+    override fun JsonNodeObject.deserializeOrThrow(): JsonNodeObject? = this
 }
